Organize app structure into packages
