+++ app/src/main/res/layout/feed_item.xml
@@ -148,7 +148,7 @@
             android:textColor="@color/colorPrimary"
             android:text="Commenter"/>
 
-        <co.adrianblan.cheddar.JellyBeanCompatTextView
+        <co.adrianblan.cheddar.views.JellyBeanCompatTextView
             android:id="@+id/feed_item_comment_text"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
