+++ app/src/main/res/layout/feed_item_alternate.xml
@@ -149,7 +149,7 @@
             android:textColor="@color/white"
             android:text="Commenter"/>
 
-        <co.adrianblan.cheddar.JellyBeanCompatTextView
+        <co.adrianblan.cheddar.views.JellyBeanCompatTextView
             android:id="@+id/feed_item_comment_text"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
